FU-123 refactor: 사진작가측 신청서관련 서비스 클래스 분리
신청서 상세보기 페이지에서 일어나는 작업들은 photographerReservationDetails에서 수행하고자 함
