+++ src/main/java/com/foru/freebe/reservation/service/PhotographerReservationService.java
@@ -10,22 +10,11 @@ import org.springframework.stereotype.Service;
 
 import com.foru.freebe.common.dto.ApiResponse;
 import com.foru.freebe.constants.SortConstants;
-import com.foru.freebe.errors.errorcode.CommonErrorCode;
-import com.foru.freebe.errors.exception.RestApiException;
-import com.foru.freebe.reservation.dto.CustomerDetails;
 import com.foru.freebe.reservation.dto.FormComponent;
-import com.foru.freebe.reservation.dto.FormDetailsViewResponse;
 import com.foru.freebe.reservation.dto.FormListViewResponse;
-import com.foru.freebe.reservation.dto.PreferredDate;
-import com.foru.freebe.reservation.dto.ReferenceImageUrls;
-import com.foru.freebe.reservation.dto.StatusHistory;
-import com.foru.freebe.reservation.entity.ReferenceImage;
 import com.foru.freebe.reservation.entity.ReservationForm;
-import com.foru.freebe.reservation.entity.ReservationHistory;
 import com.foru.freebe.reservation.entity.ReservationStatus;
-import com.foru.freebe.reservation.repository.ReferenceImageRepository;
 import com.foru.freebe.reservation.repository.ReservationFormRepository;
-import com.foru.freebe.reservation.repository.ReservationHistoryRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -33,8 +22,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class PhotographerReservationService {
     private final ReservationFormRepository reservationFormRepository;
-    private final ReservationHistoryRepository reservationHistoryRepository;
-    private final ReferenceImageRepository referenceImageRepository;
 
     public ApiResponse<List<FormListViewResponse>> getReservationList(Long photographerId) {
         List<FormListViewResponse> data = getReservationListAsStatus(photographerId);
@@ -43,32 +30,6 @@ public class PhotographerReservationService {
             .message("Successfully get reservation list")
             .status(200)
             .data(data)
-            .build();
-    }
-
-    public ApiResponse<FormDetailsViewResponse> getReservationFormDetails(Long photographerId, Long formId) {
-        ReservationForm reservationForm = findReservationForm(photographerId, formId);
-        List<StatusHistory> statusHistories = getStatusHistories(reservationForm);
-
-        CustomerDetails customerDetails = buildCustomerDetails(reservationForm);
-        Map<String, String> shootDetails = reservationForm.getPhotoInfo();
-        Map<Integer, PreferredDate> preferredDates = reservationForm.getPreferredDate();
-        ReferenceImageUrls preferredImages = getPreferredImages(reservationForm);
-
-        FormDetailsViewResponse formDetailsViewResponse = FormDetailsViewResponse.builder(reservationForm.getId(),
-                reservationForm.getReservationStatus(), statusHistories, reservationForm.getProductTitle(), customerDetails,
-                shootDetails, preferredDates)
-            .photoOptions(reservationForm.getPhotoOption())
-            .originalImage(preferredImages.getOriginalImage())
-            .thumbnailImage(preferredImages.getThumbnailImage())
-            .requestMemo(reservationForm.getCustomerMemo())
-            .photographerMemo(reservationForm.getPhotographerMemo())
-            .build();
-
-        return ApiResponse.<FormDetailsViewResponse>builder()
-            .message("Successfully get reservation list")
-            .status(200)
-            .data(formDetailsViewResponse)
             .build();
     }
 
@@ -124,50 +85,5 @@ public class PhotographerReservationService {
             reservationForm.getProductTitle(),
             reservationForm.getShootingDate() == null ? null : reservationForm.getShootingDate()
         );
-    }
-
-    private ReservationForm findReservationForm(Long photographerId, Long formId) {
-        return reservationFormRepository.findByPhotographerIdAndId(photographerId, formId)
-            .orElseThrow(() -> new RestApiException(CommonErrorCode.RESOURCE_NOT_FOUND));
-    }
-
-    private List<StatusHistory> getStatusHistories(ReservationForm reservationForm) {
-        return reservationHistoryRepository.findAllByReservationForm(reservationForm)
-            .orElseThrow(() -> new RestApiException(CommonErrorCode.RESOURCE_NOT_FOUND))
-            .stream()
-            .map(this::toStatusHistory)
-            .collect(Collectors.toList());
-    }
-
-    private StatusHistory toStatusHistory(ReservationHistory reservationHistory) {
-        return StatusHistory.builder()
-            .reservationStatus(reservationHistory.getReservationStatus())
-            .statusUpdateDate(reservationHistory.getStatusUpdateDate())
-            .build();
-    }
-
-    private CustomerDetails buildCustomerDetails(ReservationForm reservationForm) {
-        return CustomerDetails.builder()
-            .name(reservationForm.getCustomer().getName())
-            .phoneNumber(reservationForm.getCustomer().getPhoneNumber())
-            .instagramId(reservationForm.getInstagramId())
-            .build();
-    }
-
-    private ReferenceImageUrls getPreferredImages(ReservationForm reservationForm) {
-        List<ReferenceImage> referenceImages = referenceImageRepository.findAllByReservationForm(reservationForm);
-
-        List<String> originalImageUrls = referenceImages.stream()
-            .map(ReferenceImage::getOriginUrl)
-            .collect(Collectors.toList());
-
-        List<String> thumbnailImageUrls = referenceImages.stream()
-            .map(ReferenceImage::getThumbnailUrl)
-            .collect(Collectors.toList());
-
-        return ReferenceImageUrls.builder()
-            .originalImage(originalImageUrls)
-            .thumbnailImage(thumbnailImageUrls)
-            .build();
     }
 }
